✅ Add integration tests
This commit adds the ability to execute integration tests.
The purpose of such tests is to test the application
when it is "integrated" with the dependencies (i.e. database),
to verify that it is working end-to-end. Although not used
by the current integration test, the dagger function
is made to bind against the ledger service, to allow for
actual tests against the server in future commits.