👷 Remove `README.md` from api generated code
This commit removes the `README.md` from api
generated code. The `README.md` has text which
changes each time the code is generated, even
if there is no change (as the time is printed).
This leads to conflicts across PRs and also
prevents the diff between the previous and new states
of generated code, being clean. This is particularly
important because in the future, if a diff is detected
between the current state and the future state,
then it will be used to fail the PR (so that the
author is forced to rerun the generate stage).