✨ detect drift in api generated code
This commit introduces an additional check to the
check-pull-request job, which runs api spec
generation and compares it to what is currently in the
repo. The intention is to detect whether the generated
code is in anyway different to what is in the repo,
in which case the check-pull-request job should fail.
This will enforce that developers dont forget
to rerun api generation after changing the spec.