🎨 Override default openapi router init
Overrides the default openapi router initialization with a custom
one defined in the server package. Ultimately this
is required to control aspects of the server more closely.
As a result of control the router initialization, the logging
can be controlled, and in this commit is set to the zap logger
which is the desired logger for this project.